backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
databases[postgresql]==0.8.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
aiosmtplib==3.0.1
jinja2==3.1.2
email-validator==2.1.0

# Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Rate Limiting & Middleware
slowapi==0.1.9
python-redis==5.0.1

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.39.1

# Utilities
python-dotenv==1.0.0
click==8.1.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
gunicorn==21.2.0